+++ familyhelper-clannad-impl/src/test/java/com/dwarfeng/familyhelper/clannad/impl/configuration/FastJsonConfiguration.java
@@ -3,6 +3,7 @@ package com.dwarfeng.familyhelper.clannad.impl.configuration;
 import com.alibaba.fastjson.parser.ParserConfig;
 import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonGenderTypeIndicator;
 import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonProfile;
+import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonUser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Configuration;
@@ -17,6 +18,7 @@ public class FastJsonConfiguration {
         //实体对象。
         ParserConfig.getGlobalInstance().addAccept(FastJsonProfile.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonGenderTypeIndicator.class.getCanonicalName());
+        ParserConfig.getGlobalInstance().addAccept(FastJsonUser.class.getCanonicalName());
         LOGGER.debug("FastJson autotype 白名单配置完毕");
     }
 }