+++ familyhelper-clannad-node/src/main/java/com/dwarfeng/familyhelper/clannad/node/configuration/ServiceConfiguration.java
@@ -1,13 +1,16 @@
 package com.dwarfeng.familyhelper.clannad.node.configuration;
 
+import com.dwarfeng.familyhelper.clannad.impl.service.operation.UserCrudOperation;
 import com.dwarfeng.familyhelper.clannad.stack.bean.entity.GenderTypeIndicator;
 import com.dwarfeng.familyhelper.clannad.stack.bean.entity.Profile;
+import com.dwarfeng.familyhelper.clannad.stack.bean.entity.User;
 import com.dwarfeng.familyhelper.clannad.stack.cache.GenderTypeIndicatorCache;
 import com.dwarfeng.familyhelper.clannad.stack.cache.ProfileCache;
 import com.dwarfeng.familyhelper.clannad.stack.dao.GenderTypeIndicatorDao;
 import com.dwarfeng.familyhelper.clannad.stack.dao.ProfileDao;
 import com.dwarfeng.sfds.api.integration.subgrade.SnowFlakeLongIdKeyFetcher;
 import com.dwarfeng.subgrade.impl.bean.key.ExceptionKeyFetcher;
+import com.dwarfeng.subgrade.impl.service.CustomBatchCrudService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
 import com.dwarfeng.subgrade.stack.bean.key.KeyFetcher;
@@ -32,19 +35,22 @@ public class ServiceConfiguration {
     private final ProfileCache profileCache;
     private final GenderTypeIndicatorDao genderTypeIndicatorDao;
     private final GenderTypeIndicatorCache genderTypeIndicatorCache;
+    private final UserCrudOperation userCrudOperation;
 
     public ServiceConfiguration(
             ServiceExceptionMapperConfiguration serviceExceptionMapperConfiguration,
             ProfileDao profileDao,
             ProfileCache profileCache,
             GenderTypeIndicatorDao genderTypeIndicatorDao,
-            GenderTypeIndicatorCache genderTypeIndicatorCache
+            GenderTypeIndicatorCache genderTypeIndicatorCache,
+            UserCrudOperation userCrudOperation
     ) {
         this.serviceExceptionMapperConfiguration = serviceExceptionMapperConfiguration;
         this.profileDao = profileDao;
         this.profileCache = profileCache;
         this.genderTypeIndicatorDao = genderTypeIndicatorDao;
         this.genderTypeIndicatorCache = genderTypeIndicatorCache;
+        this.userCrudOperation = userCrudOperation;
     }
 
     @Value("${cache.timeout.entity.profile}")
@@ -80,6 +86,16 @@ public class ServiceConfiguration {
     public DaoOnlyEntireLookupService<GenderTypeIndicator> genderTypeIndicatorDaoOnlyEntireLookupService() {
         return new DaoOnlyEntireLookupService<>(
                 genderTypeIndicatorDao,
+                serviceExceptionMapperConfiguration.mapServiceExceptionMapper(),
+                LogLevel.WARN
+        );
+    }
+
+    @Bean
+    public CustomBatchCrudService<StringIdKey, User> userBatchCustomCrudService() {
+        return new CustomBatchCrudService<>(
+                userCrudOperation,
+                new ExceptionKeyFetcher<>(),
                 serviceExceptionMapperConfiguration.mapServiceExceptionMapper(),
                 LogLevel.WARN
         );